dev - Working on backlog communication.

+++ Tyr/utilities.h
@@ -7,5 +7,5 @@ inline float getDistance( int x1, int y1, int x2, int y2 )
 	x2 -= x1;
 	y2 -= y1;
 
-	return sqrtf( x2*x2 + y2*y2 );
+	return (float)sqrtf( x2*x2 + y2*y2 );
 }
